Handle polling of temperatures in main view
This was a very silly bug. I had every single
temperature channel poll all channel temps
which lead to O(n²) network traffic. Now, the
temperatures are polled one time in the main
view and written to a store, which can be
used by the individual channels to read from.